Factor out EmitResultTypeInit
Initializing the overall result variable is more complicated than:

    var item ResultType

The reason is because of Go's JSON marshaller that serializes nil arrays as
null rather than an empty array. Additionally, using a method could allow us
to preallocate the array based on the number of rows by adding another param.

Related to #6.