Add support for void columns
Void types commonly occur with function calls in projected columns. Add the
necessary plumbing to propagate the void type using a separate type. When
emitting Go code, we emit nil for row scan parameters but ignore void
columns in return types and in structs.

Void types complicate how we handle the ergonomic return values when there's
only 1 output column. We might have a query that has two void columns and one
int column. In that case, we return an int, not a Row.

Fixes #11.