Handle mismatch between explain outputs and expected cols
The explain plan output doesn't always match up with the SELECT columns.
If the top explain node is append, the outputs are empty because the child
nodes hold the output.

The proper solution is a full plan parser that has some knowledge about node
semantics. For now, continue to hack away with our top-level node heuristics.

Related to #13.