Generate all querier methods in a package
Fixes a bug where pggen only generates methods for the leader file. To fix,
introduce the concept of a TemplatedPackage, which contains all query files for
a single pggen invocation. Then, make the leader file responsible for iterating
over every file in the package to generate the Querier interface.

Fixes #20.