Use genericBatch in place of *pgx.Batch
I want to allow for custom implementations of batch functionality. I
specifically need it to allow for customizing the generated SQL. By implementing
genericConn, we can modify the SQL that's sent with conn.SendBatch.

For Query, QueryRow, and Exec, we can implement genericConn for similar hooks.
But since *pgx.Batch is a struct, we can't override it. By using an interface,
we can pass a custom wrapper over *pgx.Batch.