Remove all batch queries (#93)
PGXv5 dramatically changes how batch queries work. In v4 we use a separate
batch and scan call. In v5, we use a callback with generics for type safety.

Remove batch queries for now since it complicates upgrading to pgx V5. We may
them back in later with more ergonomic API.

#74